backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.1
drf-yasg>=1.21.6
djangorestframework-simplejwt>=5.3.1
django-import-export>=3.3.4
celery>=5.5.3
django-daisy>=1.1.1